requirements.txt
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-ninja==1.2.1
django-picklefield==3.2
django-q==1.3.9
django-role-permissions==3.2.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
redis==3.5.3
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
